requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

asttokens==3.0.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-mgmt-storage==22.1.1
azure-storage-blob==12.23.1
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
linecache2==1.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
msal==1.32.0
msal-extensions==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
coverage==7.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traceback2==1.4.0
traitlets==5.14.3
typing_extensions==4.12.2
unittest2==1.1.0
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
